Added the coin dice model for 2 sided dice.
For edge detection on the faces, I use the distance from the center point.

For the edges of the coin, no edge detection is done.  Instead I set the normal and corner normal to be equal to the corner normal and set the edges to be something that the point will never be near to.  Thus, the shader will iterpolate the corner normal smoothly over the edge of the coin.