Adding the ability to put any symbol on the die sides.
Added the Dice Sides activity to specify the symbol, value, and whether it will be re-rolled if hit.  Added features to try to fit text on the die side with the biggest possible font.  Added the ability to choose a die color.  Also fixed some bugs.

In the dice customization activity, if the user clicks on save and then enters a file name that contains a directory separator, then openFileOutput will throw an IllegalArgumentException.  The code only caught IOException, so it would crash.  The fix is to catch all Exception and then throw up an errror dialog with the error message.

Fixed minor display issue where it would display +-x in the equation to the left of the equals sign if the value is negative.  Now it just displays -x.

Fixed a bug where the dice side color might change if the "rainbow" color is used and there are multiple sides with the same symbol and it has to redraw the dice (the app was reloaded and the dice roll was loaded from the log file, or the surface changed and the dice roll was reloaded from the log file).

Fixed a bug where if a dice roll is reloaded from the log file and there was a constant in the dice list (but not at the end), then the result would not display properly.

performance: switched to using std::unique_ptr<unsigned char[]> instead of std::vector<char> to improve performance.  For some reason, std::vector<char> bitmap(bitmapSize); was very bad performance wise.  It took several seconds.

Fixed bug where textures don't show up if there are too many of them.  The problem is that the textures were stored in a long thin strip.  One column and many rows of textures for the dice faces.  If there are too many textures (dice sides), then it would exceed the maximum limit for textures for Vulkan and OpenGL and the textures would not display properly.  Switched to using multiple columns to place the textures instead of a single strip.  In fact, the code tries to make a square texture.

In the dice customization activity, switched to using text changed listeners instead of on focus changed listeners.  This way the text changes immediately instead of having to wait for a focus lost event.

When clicking on the new button in Dice Customization, replicate the last dice config in the list.

Made limits on lengths of all EditTexts.

Only display the result once if it is a single value.

Added recipes dialog.

Sometimes rotate text on the triangular sided dice and the 30 sided die depending on the aspect ratio.  If the aspect ratio of the text is such that the width is greater than the height on triangular sided dice and the triangle is an isosceles triangle (and not equilateral), then it will rotate the text such that the long side of the text is on the long side of the triangle.

Center text on die face and add padding.  So that the text does not run into the bottom of the die where the surface starts "curving".

Change the texture coordinates computation.  Move most of the conversion from bitmap coordinates to texture coordinates to the Java code.  The C code still needs to compute the texture coordinate for the vertices given the texture coordinates for the top and bottom of each symbol.

made changes to display emojis correctly.

Allow the user to select a color for the dice.  Allowed colors are red, orange, yellow, green, blue, purple, black, white, and rainbow.