bug fixes.
Store the number of the config being edited when saving the instance state so that when restoring from the instance state, the previous item being edited becomes the current item being edited.

For the color spinner, set the background color to the color being selected except if it is rainbow being selected.  For rainbow, do not set the color if the TextView is being created.  If it is being reused, set it to the default text color and for the background, use the color opposite of the text color.