support gestures: scroll, scale, zoom.
Add the pinch and zoom gesture.

Add recognition of the scale gesture.  The scale gesture moves the view point closer to or further away from the screen, making the dice larger or smaller.

Recognize the scroll gesture.  The scroll gesture just scrolls the screen.  Limit the amount that users can scroll.

Fix problem for some devices where x and y are swapped.  For some devices or display modes the code would draw the objects on the side of the screen when they should be drawn at the top.  Also, the objects would be skewed.  If I reversed the width and height in the perspective matrices, things would be fine.  The proper solution is to check VKSurfaceCapabilitiesKHR::currentTransform and set VkSwapchainCreateInfoKHR::preTransform to this value.  Then perform the approriate transform.  Use VK_SURFACE_TRANSFORM_IDENTITY_BIT_KHR if we get one of the mirror cases since we can't test those transforms.

Fix GL coordinate system.  Only invert the y-axis on the perspective matrix instead of doing a bunch of work to invert the texture coordinates and the positions.

Remove dead OpenGL coordinate transformations.  Since we are using the perspective matrix to invert the y-axis, we no longer need special code for OpenGL textures and coordinates.  The exception is for the 30 sided die.  We still need to invert the z-axis in the perspective matrix for the 30-sided die.  I am still not sure why this is needed and only needed for OpenGL for only this die type.

preparing for pinch and zoom gestures.  Move the projection and view matrices to the base class.  Make the screen width and height 2.0f in Vulkan/OpenGL coordinates and let the user zoom and scroll if they need to.  Move the viewPoint vector from a global file to being a member of the RainbowDice class.  Also made it non-const.  It needs to change as the user zooms the screen.

Fix to drawer going away for screen resizes while drawing.

Keep the drawer running when the dice are done rolling.  Add a worker loop which receives draw events, stop events, or surface dirty events on a dice channel.  Added a dice channel which controls the sending of events to the drawing worker.

Trust Java's values for the Window size rather than the swapchain (in Vulkan).