Refactored common code and bug fixes.
Made a template class (RainbowDiceGraphics) containing the dice list so that the dice list is either a list of vulkan dice or GL dice.  There are two classes inheriting from this template which have all of the vulkan and GL specific code.

There is also a DiceGraphics template class which has all of the common code related to a specific die.  There are also two classes inheriting from this template class that contain all the vulkan or GL only per die logic.

Fixed bug where you could not select some dice.  If the dice had y > 1.0f or y < -1.0f after rolling stopped and the dice have been moved to their stopped position, then they could not be selected.  Make sure the check for whether the dice are off the screen and adjusting x, y, z so they are back on the screen only occurs if the dice are rolling.