allow user to reroll selected dice.
Allow the user to reroll dice they have selected.  Modified the way a result is displayed from the log file.  It now sends the real dice configuration to the worker along with a vector of vector of up face indices.  Each vector from the base vector represents a die from the dice config.  If this vector has multiple values, the dice was rerolled and added in.  "modified roll" will appear in the dice results if some of the dice were rerolled (both in the log and in the results text view).