Add user ability to delete selected dice.
Also had to refactor the dice list into a list of lists.  The top level list corresponds one to one with the dice config.  The sublist contains rerolls.  More than one die in this list means that a die was rerolled for some reason (either through the automatic reroll system or because the user selected to add them).  Zero items in the sublist indicates that the die was deleted.  1 item means that no add and reroll occurred.