Improve shutdown code, remove dead code.
The render thread should be stopped if onPause is called not just if SurfaceHolder.Callback.surfaceDestroyed is called.  Now the render thread is stopped in onPause or SurfaceHolder.Callback.surfaceDestroyed, which ever happens first.

Removed some dead code.

Improved shutdown time by making the worker thread return a void, not a java string when it exits the C++ code.  If an error occurs in the worker thread, then the worker thread uses the Notify object to send the error back to java before shutting down.  This way it only takes longer to shutdown for the case of an error occurring.