Display graphics API and device in Settings Screen
Added plumbing to get the informaton on which graphics API is being used: Vulkan or OpenGL ES.  If it is Vulkan, then the version of Vulkan API and the device name are also obtained.

Changed the "Theme Selection" screen into a "Settings" screen.

Report the information to the user in the "Settings" screen.