Correct screen placement of stopped dice.
Use the inverse of the perspective matrix and the inverse of the original view matrix (before the user changes it) to calculate world coordinates from screen coordinates.  Use these coordinates to place the dice on the screen.

Made the maximum x and y in DicePhysicsModel be set according to what the max x and y are in world space at the max z for a particular perspective matrix.