Fix swapchain image size on resize
Pass the values for width and height from Java to the swapchain constructor then if the width and height fall in between minImageExtent and maxImageExtent, then use the width and height from Java for the imageExtent when recreating the swapchain.

Also, draw a frame before recreating the swapchain.  This seems to fix the problem.  However, it is unclear for how many cases and why this works at all.