Add render options
Added an option to allow a user to select to use OpenGL instead of Vulkan.

Added an option to allow a user to choose not to use gravity and instead have gravity point towards the screen always.  Needs work.

Added an option to allow a user to choose whether to animate the dice rolling or just draw the stopped dice.