Fix the use gravity flag.
Add use of Linear Acceleration and Gravity sensors.  Prefer the use of the Linear Acceleration and Gravity sensors over the use of the accelerometer.  If those sensors are not present, use the accelerometer.  If the use gravity flag is selected, then the gravity sensor is not needed.  Force the use of gravity if only the accelerometer is present.

also fixed bugs:

Avoid having the dice jump in location if the screen is resized during a roll.  If the die moves too much, just use the previous position and return false (does not need redraw).

Set the TextView text size to 20sp for the dice result text box, so that it is more visible.

Set the minimum size to 50sp for the dice result TextView.  This is done so that the dice don't move immediately after a roll.  The text for the result is often two lines with the new font size.  It is distracting and confusing to have the dice shift position after a roll.