bug fixes
Changed the color of d6-d6(re-roll 6's).

Fixed bug that occurred if there was nothing to render and the screen was resized.  This would fail in drawFrame because the command buffers are not initialized if there is nothing to draw.  Only draw a frame on resize if there are dice to draw.