minor fixes
call Surface.release() when we are done with the surface in case this helps with the bug where if onPause does not finish before the app is placed in the foreground again, the old surface remains as a background to the new surface.  It does not seem to fix the bug in android 8.0.

Removed catching of ... in Notify::sendResult.  If any exceptions occur in there, something really awful must have occurred.  We don't expect to get exceptions in there.  So we catch the exception at the JNI call to start the worker, then we send the exception to java, then we exit the worker.