fixed bugs
Bug were some of the dice would get stuck if window was resized while drawing occurred.  I think they got stuck outside the area of that they were allowed to bounce around in.  Then they could not ever get back in because it would always cause a jump in position that was too big.

Fixed result display where if the dice were using symbols with no values and each die rolled a different symbol.  In this case it would list the result = result in a different order.  For this case, the result should only show once and in the order that the dice are displayed.

Reset the "reset view" button to disabled and set the image to grayed out.