Bug fixes.
Free jobject that are allocated in a loop.  We do these even for short lived JNI calls because the loops could create a large number of jobject references that do not get freed until the end of the call.  The Android documentation states that it is only required to reserve 16 slots for local references and that you should free as you go if you need more.

For consistancy: when the user rerolls or adds a die, and its result is a symbol that requires automatic reroll, automatically reroll the die.  This was the behavior for animating rolling dice, this change extends this behavior to just drawing the die result.