Fixed crash bugs.
Don't use std::move on trying Vulkan in case an exception is thrown in loading Vulkan.  In that case it will try OpenGL and we need to keep around the pointer for this case.

Don't call RainbowDice::drawFrame if there are no dice present.  This causes a crash.

Incremented version for release.