bug fix: uninitialized model matrix.
When selecting to not animate the dice rolling (and just display a result), the model matrix was not initialized before calling randomizeUpFace.  randomizeUpFace calls getAngleAxis to get the angle and axis to compute the quaternion to make the randomly selected face point towards the screen.  If the uninitialized model matrix turns out to be all zeros, then getAngleAxis will set the angle and axis to NaN for all values as the model matrix is used for these calculations and must not be all zeros.  Setting the model matrix to the values obtained from calling glm::scale on the a vector containing the radii of the die before the call to randomizeUpFace fixes this issue.