Let users to select a bonus to apply to all rolls.
Add a menu item to the main screen stating the current bonus to add to all rolls.  If this menu item is clicked on, throw up a dialog to let the user select a new bonus that will either apply to the current roll only, all future rolls, or both the current roll and all future rolls.  The dialog also has a cancel button.

If neither current roll nor future rolls is checked, disable the "OK" button and gray out the text.