allow users to select the direction of gravity.
Users can now select for gravity to go down or up.  (Up used to be the direction the dice "fell".)  When dice are falling down, a box for them to roll in is shown.  When the dice stop rolling and are all moved to the top of the screen, then the box is removed.  The dice box uses a different shader (and thus a different pipeline in Vulkan).  Users can select the direction for gravity in the settings activity.

Fix Vulkan validation.