Move stopped dice to below where they are rolling.
Move dice that are not rolling to below the surface of rolling dice.  This way they are out of the way view wise.  Make them larger so that they are move visible.

Change edge thickness depending on size of die.  Pass the edge thickness into the shader and allow different dice to have different edge thickness and change the edge thickness when the dice are stopped.